requirements.txt
tensorflow>=2.15.0
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
numpy>=1.24.0
pandas>=2.0.0
streamlit>=1.28.0
medmnist>=1.2.0
pillow>=10.0.0
seaborn>=0.12.0 